requirements.txt
tkinter
paho-mqtt
PySimpleGUI
pytest